feat: Implement folder animation and layout adjustments (test-folder-anim)
This commit introduces animations for folder opening and closing,
along with adjustments to the folder grid layout.

Key changes include:

- **Folder Animation:**
  - Implemented `animatedRect2` to define the animated bounds of the
    folder grid during transitions.
  - Modified `FolderScreen` to use `animatedRect2` for sizing and
    positioning the `FolderGridLayout`.
  - Added `getAnimatedRect2` function to calculate the animated
    rectangle based on animation progress and layout information.
  - Updated `FolderPopupLayoutInfo` to store initial test dimensions
    and center coordinates for animation calculations.

- **Folder Grid Layout:**
  - Removed `layoutWidth` and `layoutHeight` parameters from
    `FolderGridLayout` as they are no longer needed.
  - The `FolderGridLayout` now calculates cell dimensions based on the
    available constraints and the number of columns/rows.
  - Adjusted `FolderPopupLayoutInfo` to store `startTestWidthPx` and
    `startTestHeightPx` for initial animation state.

These changes enhance the user experience by providing smooth visual
feedback during folder interactions and improve the flexibility
of the folder grid layout.
