+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/component/FolderGrid.kt
@@ -1,0 +1,149 @@
+package com.eblan.launcher.feature.home.component
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxScope
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ParentDataModifier
+import androidx.compose.ui.layout.SubcomposeLayout
+import androidx.compose.ui.unit.Constraints
+import androidx.compose.ui.unit.Density
+import com.eblan.launcher.domain.model.GridItem
+import com.eblan.launcher.feature.home.util.FOLDER_PREVIEW_COLUMNS
+import com.eblan.launcher.feature.home.util.FOLDER_PREVIEW_ROWS
+
+@Composable
+internal fun PreviewFolderGridLayout(
+    modifier: Modifier = Modifier,
+    gridItems: List<GridItem>?,
+    previewColumns: Int = FOLDER_PREVIEW_COLUMNS,
+    previewRows: Int = FOLDER_PREVIEW_ROWS,
+    content: @Composable BoxScope.(GridItem) -> Unit,
+) {
+    SubcomposeLayout(modifier = modifier) { constraints ->
+        val previewCellSize = minOf(
+            constraints.maxWidth,
+            constraints.maxHeight,
+        ) / maxOf(
+            previewColumns,
+            previewRows,
+        )
+
+        val previewGridWidth = previewCellSize * previewColumns
+
+        val previewGridHeight = previewCellSize * previewRows
+
+        val previewOffsetX = (constraints.maxWidth - previewGridWidth) / 2
+
+        val previewOffsetY = (constraints.maxHeight - previewGridHeight) / 2
+
+        layout(
+            width = constraints.maxWidth,
+            height = constraints.maxHeight,
+        ) {
+            gridItems
+                ?.take(previewColumns * previewRows)
+                ?.forEachIndexed { index, gridItem ->
+                    subcompose(gridItem.id) {
+                        val x = previewOffsetX + (index % previewColumns) * previewCellSize
+
+                        val y = previewOffsetY + (index / previewColumns) * previewCellSize
+
+                        Box(
+                            modifier = Modifier.folderGridItem(
+                                x = x,
+                                y = y,
+                                width = previewCellSize,
+                                height = previewCellSize,
+                            ),
+                        ) {
+                            content(gridItem)
+                        }
+                    }.forEach { measurable ->
+                        val parentData = measurable.parentData as FolderGridItemParentData
+
+                        measurable.measure(
+                            Constraints.fixed(
+                                width = parentData.width,
+                                height = parentData.height,
+                            ),
+                        ).placeRelative(
+                            x = parentData.x,
+                            y = parentData.y,
+                        )
+                    }
+                }
+        }
+    }
+}
+
+@Composable
+internal fun FolderGridLayout(
+    modifier: Modifier = Modifier,
+    gridItems: List<GridItem>?,
+    columns: Int,
+    rows: Int,
+    width: Int,
+    height: Int,
+    content: @Composable BoxScope.(GridItem) -> Unit,
+) {
+    SubcomposeLayout(modifier = modifier) { constraints ->
+        val cellWidth = width / columns
+        val cellHeight = height / rows
+
+        layout(
+            width = constraints.maxWidth,
+            height = constraints.maxHeight,
+        ) {
+            gridItems?.forEachIndexed { index, gridItem ->
+                subcompose(gridItem.id) {
+                    Box(
+                        modifier = Modifier.folderGridItem(
+                            x = (index % columns) * cellWidth,
+                            y = (index / columns) * cellHeight,
+                            width = cellWidth,
+                            height = cellHeight,
+                        ),
+                    ) {
+                        content(gridItem)
+                    }
+                }.forEach { measurable ->
+                    val parentData = measurable.parentData as FolderGridItemParentData
+
+                    measurable.measure(
+                        Constraints.fixed(
+                            width = parentData.width,
+                            height = parentData.height,
+                        ),
+                    ).placeRelativeWithLayer(
+                        x = parentData.x,
+                        y = parentData.y,
+                    )
+                }
+            }
+        }
+    }
+}
+
+private fun Modifier.folderGridItem(
+    x: Int,
+    y: Int,
+    width: Int,
+    height: Int,
+) = then(
+    object : ParentDataModifier {
+        override fun Density.modifyParentData(parentData: Any?) = FolderGridItemParentData(
+            x = x,
+            y = y,
+            width = width,
+            height = height,
+        )
+    },
+)
+
+private data class FolderGridItemParentData(
+    val x: Int,
+    val y: Int,
+    val width: Int,
+    val height: Int,
+)
